+++ src/cli/generateOne.ts
@@ -1,5 +1,6 @@
-import { command, flag, restPositionals, string } from 'cmd-ts';
-import { glob } from 'node:fs/promises';
+import { command, flag, restPositionals } from 'cmd-ts';
+import { getFiles } from '../functions';
+import { path } from './custom';
 import { watcher } from './helpers';
 
 export const generateOne = command({
@@ -17,33 +18,15 @@ export const generateOne = command({
     files: restPositionals({
       description: 'The files to generate',
       displayName: 'Files',
-      type: string,
-    }),
-
-    strict: flag({
-      description: 'Enable strict mode',
-      short: 's',
-      env: 'APP_TYPINGS_STRICT_ONE',
-      long: 'strict',
+      type: path,
     }),
   },
-  handler: async ({ watch: persistent, files, strict }) => {
+  handler: async ({ watch: persistent, files }) => {
     const isEmpty = files.length === 0;
     if (isEmpty) return console.warn('No files specified for generation.');
 
-    const FILES = await Array.fromAsync(glob(files));
-
+    const FILES = await getFiles(...files);
     if (FILES.length === 0) return console.warn('Files not found');
-
-    const hasMachineNames = FILES.every(file =>
-      file.endsWith('.machine.ts'),
-    );
-
-    if (strict && !hasMachineNames) {
-      return console.warn(
-        'Strict, all machines names must ends with ".machine.ts"',
-      );
-    }
 
     return watcher(persistent, ...FILES);
   },